Models/Vision_model/Fastapi_app/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
transformers
python-multipart
python-dotenv
groq
requests